+++ benches/bench_io.rs
@@ -80,6 +80,7 @@ fn bench_ms_init_mwax_half_1247842824(crt: &mut Criterion) {
                     obs_ctx.array_pos,
                     obs_ctx.ant_positions_geodetic().collect(),
                     Duration::from_total_nanoseconds(0),
+                    true,
                 );
                 ms_writer
                     .initialize_mwa(
@@ -136,6 +137,7 @@ fn bench_uvfits_init_mwax_half_1247842824(crt: &mut Criterion) {
                     obs_ctx.name.as_deref(),
                     obs_ctx.ant_names.clone(),
                     obs_ctx.ant_positions_geodetic().collect(),
+                    true,
                     None,
                 )
                 .unwrap();
@@ -208,6 +210,7 @@ fn bench_ms_write_mwax_part_1247842824(crt: &mut Criterion) {
                     obs_ctx.array_pos,
                     obs_ctx.ant_positions_geodetic().collect(),
                     Duration::from_total_nanoseconds(0),
+                    true,
                 );
                 ms_writer
                     .initialize_mwa(
@@ -266,6 +269,7 @@ fn bench_uvfits_write_mwax_part_1247842824(crt: &mut Criterion) {
                     obs_ctx.name.as_deref(),
                     obs_ctx.ant_names.clone(),
                     obs_ctx.ant_positions_geodetic().collect(),
+                    true,
                     None,
                 )
                 .unwrap();